🚚 Separate method for calling chat page from AiApiController
AiApiController 에서 단순 채팅창을 조회하는 메서드는 web 패키지로 이동
